Removed collision_class in favor of class methods returning junctions.
Now instead of using if self.collision_class(Brick) -> $brick, you use if
self.collision(Brick._any) -> $brick and it magically removes any non-colliding
objects from the junction.  _any was used instead of any because of a bug that
causes infinite recursion for some reason.  It will be changed if this bug is
fixed.  examples/breakout does not yet use this yet because it is too slow; it
checks collisions on the bricks' step instead.