Rewrote sprite registration and implemented per-sprite events.
This is a bit slower, but it will be worth it in the long run.  Sprites can
now have events called on them at appropriate times, although currently just
step and draw events are implemented that mirror the old step and draw methods.
I also prettified the bricks a bit with an image file overlaid on their color.